Moved the app and app unit tests to be part of the main solution
