+++ App/App/App.csproj
@@ -6,17 +6,17 @@
 		<!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
 		<!-- <TargetFrameworks>$(TargetFrameworks);net7.0-tizen</TargetFrameworks> -->
 		<OutputType>Exe</OutputType>
-		<RootNamespace>RoboticaApp</RootNamespace>
+		<RootNamespace>App</RootNamespace>
 		<UseMaui>true</UseMaui>
 		<SingleProject>true</SingleProject>
 		<ImplicitUsings>enable</ImplicitUsings>
 
 		<!-- Display name -->
-		<ApplicationTitle>RoboticaApp</ApplicationTitle>
+		<ApplicationTitle>App</ApplicationTitle>
 
 		<!-- App Identifier -->
-		<ApplicationId>com.companyname.roboticaapp</ApplicationId>
-		<ApplicationIdGuid>68f974df-3288-40c2-809f-453801b22687</ApplicationIdGuid>
+		<ApplicationId>com.companyname.app</ApplicationId>
+		<ApplicationIdGuid>39f70845-88c6-4e74-bf90-6eb2fe2b8122</ApplicationIdGuid>
 
 		<!-- Versions -->
 		<ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
