+++ App/App/AppShell.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="RoboticaApp.AppShell"
+    x:Class="App.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:RoboticaApp"
+    xmlns:local="clr-namespace:App"
     Shell.FlyoutBehavior="Disabled">
 
     <ShellContent
