+++ App/App/AppShell.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace App
+{
+    public partial class AppShell : Shell
+    {
+        public AppShell()
+        {
+            InitializeComponent();
+        }
+    }
+}
