+++ App/App/Platforms/Android/AndroidManifest.xml
