+++ App/App/Platforms/Windows/App.xaml
@@ -1,8 +1,8 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="RoboticaApp.WinUI.App"
+    x:Class="App.WinUI.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:RoboticaApp.WinUI">
+    xmlns:local="using:App.WinUI">
 
 </maui:MauiWinUIApplication>
