+++ App/App/Platforms/iOS/AppDelegate.cs
@@ -1,0 +1,10 @@
+﻿using Foundation;
+
+namespace App
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : MauiUIApplicationDelegate
+    {
+        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    }
+}
