+++ App/App/Platforms/iOS/Info.plist
