+++ App/App/Resources/Raw/AboutAssets.txt
