--- App/RoboticaApp/RoboticaApp.sln
@@ -1,33 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31611.283
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RoboticaApp", "RoboticaApp\RoboticaApp.csproj", "{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AppTests", "AppTests\AppTests.csproj", "{CBA63671-DC4F-4252-8B7A-15D60AA8EC34}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C7CEEFE6-5395-4FF5-A20E-F68E27ED50BF}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{CBA63671-DC4F-4252-8B7A-15D60AA8EC34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CBA63671-DC4F-4252-8B7A-15D60AA8EC34}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CBA63671-DC4F-4252-8B7A-15D60AA8EC34}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CBA63671-DC4F-4252-8B7A-15D60AA8EC34}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {61F7FB11-1E47-470C-91E2-47F8143E1572}
-	EndGlobalSection
-EndGlobal
