--- App/RoboticaApp/RoboticaApp/App.xaml.cs
@@ -1,11 +1,0 @@
-﻿namespace RoboticaApp;
-
-public partial class App : Application
-{
-	public App()
-	{
-		InitializeComponent();
-
-		MainPage = new AppShell();
-	}
-}
