--- App/RoboticaApp/RoboticaApp/AppShell.xaml.cs
@@ -1,9 +1,0 @@
-﻿namespace RoboticaApp;
-
-public partial class AppShell : Shell
-{
-	public AppShell()
-	{
-		InitializeComponent();
-	}
-}
