--- App/RoboticaApp/RoboticaApp/Platforms/Android/MainActivity.cs
@@ -1,10 +1,0 @@
-﻿using Android.App;
-using Android.Content.PM;
-using Android.OS;
-
-namespace RoboticaApp;
-
-[Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
-public class MainActivity : MauiAppCompatActivity
-{
-}
