--- App/RoboticaApp/RoboticaApp/Platforms/MacCatalyst/AppDelegate.cs
@@ -1,9 +1,0 @@
-﻿using Foundation;
-
-namespace RoboticaApp;
-
-[Register("AppDelegate")]
-public class AppDelegate : MauiUIApplicationDelegate
-{
-	protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
-}
