--- App/RoboticaApp/RoboticaApp/Platforms/Tizen/Main.cs
@@ -1,16 +1,0 @@
-using System;
-using Microsoft.Maui;
-using Microsoft.Maui.Hosting;
-
-namespace RoboticaApp;
-
-class Program : MauiApplication
-{
-	protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
-
-	static void Main(string[] args)
-	{
-		var app = new Program();
-		app.Run(args);
-	}
-}
