+++ Robotics.sln
@@ -5,22 +5,62 @@ VisualStudioVersion = 17.5.33424.131
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Robotics", "Robotics\Robotics.vcxproj", "{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App", "App\App\App.csproj", "{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AppTests", "App\AppTests\AppTests.csproj", "{F7BAC05A-4738-4C56-B58C-2921D277D158}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|Any CPU.Build.0 = Debug|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|x64.ActiveCfg = Debug|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|x64.Build.0 = Debug|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|x86.ActiveCfg = Debug|Win32
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Debug|x86.Build.0 = Debug|Win32
+		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|Any CPU.ActiveCfg = Release|x64
+		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|Any CPU.Build.0 = Release|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|x64.ActiveCfg = Release|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|x64.Build.0 = Release|x64
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|x86.ActiveCfg = Release|Win32
 		{86AB8603-BA1A-4F95-ABD4-9A66275A18F9}.Release|x86.Build.0 = Release|Win32
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x64.Build.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x64.Deploy.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x86.Build.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Debug|x86.Deploy.0 = Debug|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x64.ActiveCfg = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x64.Build.0 = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x64.Deploy.0 = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x86.ActiveCfg = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x86.Build.0 = Release|Any CPU
+		{FFE3BA5F-35BF-4947-80CB-D1FD4BAA5609}.Release|x86.Deploy.0 = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|x64.Build.0 = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Debug|x86.Build.0 = Debug|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|x64.ActiveCfg = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|x64.Build.0 = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|x86.ActiveCfg = Release|Any CPU
+		{F7BAC05A-4738-4C56-B58C-2921D277D158}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
